Increasing modularity of API
