+++ models/department.js
@@ -24,6 +24,17 @@ var departmentSchema = new mongoose.Schema({
   }
 })
 
+// Return all of the departments
+departmentSchema.statics.getAll = function (callback) {
+  this.find().sort({_id: 1}).exec(function (err, departments) {
+    if (err) {
+      console.log(err)
+    } else {
+      callback(departments)
+    }
+  })
+}
+
 // Create the Department model from the courseSchema
 var Department = mongoose.model('Department', departmentSchema)
