+++ models/semester.js
@@ -18,6 +18,17 @@ var semesterSchema = new mongoose.Schema({
   }
 })
 
+// Return all of the departments
+semesterSchema.statics.getAll = function (callback) {
+  this.find().sort({_id: -1}).exec(function (err, semesters) {
+    if (err) {
+      console.log(err)
+    } else {
+      callback(semesters)
+    }
+  })
+}
+
 // Create the Semester model from the courseSchema
 var Semester = mongoose.model('Semester', semesterSchema)
